Fix debug tracers on Python 3 with PostgreSQL
Storm >= 0.21 wraps the raw connection in a `ConnectionWrapper` object
for the purpose of exception wrapping, so we need to handle that.  (This
is a bit ugly, but `Adapter.adapt` was already poking around in private
attributes.)

Furthermore, on Python 3, psycopg2's `getquoted` returns bytes, but
`replace_placeholders` expects tokens to be text.  Cope with this.  It
would be nice if this were tested properly, but that involves setting up
the test suite to run under PostgreSQL, which is a somewhat more
involved task.